update file 6

+++ python_variable_annotations/6-sum_mixed_list.py
@@ -2,6 +2,7 @@
 """Function sum a list of floats and integers and return their sum as a float."""
 from typing import List, Union
 
-def sum_mixed_list(mxd_list: List[Union[float, int]]) -> float:
+
+def sum_mixed_list(mxd_list: List[Union[int, float]]) -> float:
     """Returns the sum of a list of floats and integers."""
     return sum(mxd_list)
